services/yara-scanner/requirements.txt
yara-python==4.3.1
kafka-python==2.0.2
watchdog==3.0.0
